fix/operation: withdrawal-unstaking integration test tmp (#246)
* impl unstaking-new state, fixed few accountings, stuck in tx size during unstaking execute

* reduced tx size by alt update

* tools: include pricing sources in ALT

* fix(reward): change accrual_rate to 2byte

* chore(migration): remove old migrations

* chore(normalization): small, minor changes

* feat(operator): complete operator cycle test

---------

+++ programs/restaking/src/modules/reward/reward_account.rs
@@ -218,7 +218,7 @@ impl RewardAccount {
         &mut self,
         receipt_token_mint: Pubkey,
         amount: u64,
-        contribution_accrual_rate: Option<u8>,
+        contribution_accrual_rate: Option<u16>,
         from: Option<&mut UserRewardAccount>,
         to: Option<&mut UserRewardAccount>,
         current_slot: u64,
